完成检验模块
基本完成代理池的逻辑，TODO：
1、 对外暴露http接口
2、 定时任务

+++ proxypool/common/types.go
@@ -29,7 +29,7 @@ type IPStorger interface {
 	// api获取代理
 	GetIp() string
 
-	GetSomeIp(start float64, end float64) []string
+	GetSomeIp(start int64, end int64) []string
 }
 
 // IPDetecter 是测试模块的接口
